requirements.txt
Flask-Migrate
Flask-Session
Flask-UUID
mysqlclient
patool
raven[flask]
validators
uuid
